requirements.txt
pytest==7.4.2
selenium==4.15.2
fpdf2==2.7.6  # Para generar PDFs si lo necesitas
pytest-html==3.2.0  # Para generar reportes HTML